+++ src/front/asm/assemble_instruction.cpp
@@ -316,6 +316,10 @@ auto assemble_instruction(
         assemble_three_register_op<&Program::opand>(program, tokens, i);
     } else if (tokens.at(i) == "or") {
         assemble_three_register_op<&Program::opor>(program, tokens, i);
+    } else if (tokens.at(i) == "bits_of_integer") {
+        assemble_double_register_op<&Program::opbits_of_integer>(program, tokens, i);
+    } else if (tokens.at(i) == "integer_of_bits") {
+        assemble_double_register_op<&Program::opinteger_of_bits>(program, tokens, i);
     } else if (tokens.at(i) == "bits") {
         assemble_op_bits(program, tokens, i);
     } else if (tokens.at(i) == "bitand") {